+++ lib/series/series.py
@@ -3,15 +3,17 @@ import time
 
 from enodo.jobs import JOB_TYPE_BASE_SERIES_ANALYSIS, JOB_STATUS_NONE, \
     JOB_STATUS_DONE
-from enodo.model.config.series import SeriesConfigModel, SeriesState
-from lib.modulemanager import EnodoModuleManager
+from enodo.model.config.series import SeriesConfigModel, SeriesState, \
+    SeriesJobConfigModel
+
+from lib.socket.clientmanager import ClientManager
 
 
 class Series:
     __slots__ = ('rid', 'name', 'series_config', 'state',
                  '_datapoint_count_lock', 'series_characteristics')
 
-    def __init__(self, name, config, state=None,
+    def __init__(self, name: str, config: dict, state=None,
                  series_characteristics=None, **kwargs):
         self.rid = name
         self.name = name
@@ -22,7 +24,7 @@ class Series:
         self.series_characteristics = series_characteristics
         self._datapoint_count_lock = asyncio.Lock()
 
-    def get_errors(self):
+    def get_errors(self) -> list:
         # To stop circular import
         from ..jobmanager import EnodoJobManager
         errors = [
@@ -31,42 +33,42 @@ class Series:
                 self.name)]
         return errors
 
-    def is_ignored(self):
+    def is_ignored(self) -> bool:
         # To stop circular import
         from ..jobmanager import EnodoJobManager
         return EnodoJobManager.has_series_failed_jobs(self.name)
 
-    async def get_module(self, job_name):
+    async def get_module(self, job_name: str) -> SeriesJobConfigModel:
         return self.series_config.get_config_for_job(job_name).module
 
-    async def get_job_status(self, job_config_name):
+    async def get_job_status(self, job_config_name: str) -> int:
         return self.state.get_job_status(job_config_name)
 
-    async def set_job_status(self, config_name, status):
+    async def set_job_status(self, config_name: str, status: int):
         self.state.set_job_status(config_name, status)
 
     @property
-    def base_analysis_job(self):
+    def base_analysis_job(self) -> SeriesJobConfigModel:
         job_config = self.series_config.get_config_for_job_type(
             JOB_TYPE_BASE_SERIES_ANALYSIS)
         if job_config is None or job_config is False:
             return False
         return job_config
 
-    def get_job(self, job_config_name):
+    def get_job(self, job_config_name: str) -> SeriesJobConfigModel:
         return self.series_config.get_config_for_job(job_config_name)
 
-    def base_analysis_status(self):
+    def base_analysis_status(self) -> int:
         job_config = self.series_config.get_config_for_job_type(
             JOB_TYPE_BASE_SERIES_ANALYSIS)
         if job_config is None or job_config is False:
             return False
         return self.state.get_job_status(job_config.config_name)
 
-    def get_datapoints_count(self):
+    def get_datapoints_count(self) -> int:
         return self.state.datapoint_count
 
-    async def add_to_datapoints_count(self, add_to_count):
+    async def add_to_datapoints_count(self, add_to_count: int):
         """
         Add value to existing value of data points counter
         :param add_to_count:
@@ -75,7 +77,7 @@ class Series:
         async with self._datapoint_count_lock:
             self.state.datapoint_count += add_to_count
 
-    async def schedule_job(self, job_config_name, initial=False):
+    async def schedule_job(self, job_config_name: str, initial=False):
         job_config = self.series_config.get_config_for_job(
             job_config_name)
         if job_config is None:
@@ -105,14 +107,14 @@ class Series:
             job_schedule['value'] = next_value
             self.state.set_job_schedule(job_config_name, job_schedule)
 
-    async def is_job_due(self, job_config_name):
+    async def is_job_due(self, job_config_name: str) -> bool:
         job_status = self.state.get_job_status(job_config_name)
         job_schedule = self.state.get_job_schedule(job_config_name)
 
         if job_status in [JOB_STATUS_NONE, JOB_STATUS_DONE]:
             job_config = self.series_config.get_config_for_job(
                 job_config_name)
-            module = await EnodoModuleManager.get_module(job_config.module)
+            module = ClientManager.get_module(job_config.module)
             if module is None:
                 self.state.set_job_check_status(
                     job_config_name,
@@ -143,14 +145,14 @@ class Series:
             "Already active")
         return False
 
-    def update(self, data):
+    def update(self, data: dict) -> bool:
         config = data.get('config')
         if config is not None:
             self.series_config = SeriesConfigModel(**config)
 
         return True
 
-    def to_dict(self, static_only=False):
+    def to_dict(self, static_only=False) -> dict:
         if static_only:
             return {
                 'name': self.name,
@@ -169,5 +171,5 @@ class Series:
         }
 
     @classmethod
-    def from_dict(cls, data_dict):
+    def from_dict(cls, data_dict: dict) -> 'Series':
         return Series(**data_dict)